fix(room): join table column types & rename field

+++ app/src/main/java/tech/salvatore/livro_android_kotlin_paulo_salvatore/model/source/local/db/entity/CreatureUserCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["creatureId", "userId"])
 data class CreatureUserCrossRef(
-    val creatureId: Long,
-    val userId: Long
+    val creatureId: Int,
+    val userId: Int
 )
